requirement.txt
langchain
langchain-google-genai
langchain_experimental
mysql-connector-python-rf
flask